Fix float-typed ldsStartOffsetB causing garbage B matrix LR offsets on GPU
globalSubtileGrid returns floats, which propagated to ldsStartOffsetB.
The float literal (e.g. 32768.0 = 0x47000000) was emitted into v_add_u32
as IEEE 754 bits instead of the integer 0x8000, corrupting B offsets.

+++ projects/hipblaslt/tensilelite/Tensile/Tests/unit/gpu_test_helpers.py
@@ -177,7 +177,7 @@ def create_writer(cfg, mi_wave_group=None):
     readSize = 2 * tileInfoA.subtileSize
     numASubtiles = tileInfoA.globalSubtileGrid[0] * tileInfoA.globalSubtileGrid[1]
     writer.ldsStartOffsetA = 0
-    writer.ldsStartOffsetB = ((numASubtiles * tileInfoA.subtileSize + readSize-1) // readSize) * readSize
+    writer.ldsStartOffsetB = int(((numASubtiles * tileInfoA.subtileSize + readSize-1) // readSize) * readSize)
 
     return writer, kernel, tileInfoA, tileInfoB
